Ajout chargement cartes + création cartes par l'utilisateur

+++ src/Headers/Cartes/Attaque.h
@@ -11,7 +11,7 @@
 //Hérite de Carte : nom et description
 class Attaque : public Carte{
     private:
-        short m_degat;
+        int m_degat;
         //Nombre de point demandés d'énergie pour lancer l'attaque
         std::vector<std::vector<short> > m_nbPoints;
     public:
